[Update] hp 자료형 통일

+++ ProjectFreeGenre/Assets/Script/Interface/IDamage.cs
@@ -1,4 +1,4 @@
 public interface IDamage
 {
-    void Attacked(int damage);
+    void Attacked(float damage);
 }
